+++ src/network/protocol.cpp
@@ -53,7 +53,7 @@ Protocol::~Protocol()
 /** Returns a network string with the given type.
  *  \capacity Default preallocated size for the message.
  */
-NetworkString* Protocol::getNetworkString(int capacity)
+NetworkString* Protocol::getNetworkString(size_t capacity)
 {
     return new NetworkString(m_type, capacity);
 }   // getNetworkString
